Fix: Use bf16 instead of fp16 for 4-bit quantization
fp16 causes NotImplementedError with BFloat16 quantization.
bf16 is compatible with 4-bit nf4 quantization.

+++ scripts/train_dpo.py
@@ -119,7 +119,7 @@ training_args = DPOConfig(
     max_length=512,
     logging_steps=5,
     save_strategy="epoch",
-    fp16=True,
+    bf16=True,  # Use bf16 instead of fp16 for compatibility with 4-bit quantization
     remove_unused_columns=False,
     report_to="none"
 )
